Add IPv6 support
Can be invoked as:
  firehol.sh -4 [whatever]  -- IPv4 mode (should show original behaviour)
  firehol.sh -6 [whatever]  -- IPv6 mode (ip6tables only)
  firehol.sh [whatever]     -- "Both" mode

In "both" mode, individual iptables_cmd calls are checked to see if they
should be applied to IPv4 or IPv6 or both.

In "both" mode iptables_save_cmd combines the output of both iptables-save
and ip6tables-save.

In "both" iptables_restore_cmd splits apart the output of iptables_save_cmd
and directs it to the appropriate restore commands.

Module loading and clean/initialise behaviours conditionally operate
various commands according to the mode.