Add ipv6neigh and ipv6router complex rules
Use of the neigbour and router rules sets up appropriate stateless
icmpv6 filters chains which should allow ipv6 auto-configuration to
work and reduces the need for the workaround in the readme.

Added a really simple example so a complete working firewall can
be seen without needing to extract all the information from the
readme.